autotest: isolate ESRIJson tests in a separate file
